+++ telers/src/types/message.rs
@@ -5125,7 +5125,7 @@ impl Message {
 
     /// Helper method for nested field `icon_color`.
     #[must_use]
-    pub fn icon_color(&self) -> Option<i64> {
+    pub fn icon_color(&self) -> Option<i32> {
         match self {
             Self::ForumTopicCreated(val) => {
                 let inner = &val.forum_topic_created;
